requirements.txt
streamlit>=1.30.0
pandas>=2.1.0
numpy>=1.26.0
plotly>=5.18.0
scikit-learn>=1.4.0
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-huggingface>=0.1.0
langchain-community>=0.3.0
huggingface_hub>=0.20.0
python-docx>=1.1.0
google-api-python-client>=2.0.0
requests>=2.31.0
